==> R/03_describe.qmd <==
---
title: "03_describe"
format:
  html:
    embed-resources: true
editor: visual
---

# Load data

```{r}
clinical_data <- read_tsv(
  file = "../data/clinical_data_clean.tsv",
  na = 'NA')
```

# Clinical data overview

## Statistical summary of age distribution and survival analysis by cancer type

```{r}
stats_table <-
  clinical_data |>   
  mutate(age_group = cut(age_at_diagnosis,
                         breaks = c(20, 30, 40, 50, 60, 70, 80))) |>
  mutate(age_group = factor(age_group),
         cancer_type_detailed = factor(cancer_type_detailed)) |> 
  filter(cancer_type_detailed != 'NA') |> 
  table1(x = formula(~ age_group + 
                       overall_survival + 
                       overall_survival_months 
                     | cancer_type_detailed),
         data = _)
stats_table
```

## Visualizing the age distribution

```{r}
age_distribution <- 
  clinical_data |> 
  mutate(age_group = cut(age_at_diagnosis,
                         breaks = c(20, 30, 40, 50, 60, 70, 80, 100))) |> 
  group_by(age_group) |> 
  summarise(n = n()) |> 
  ggplot(aes(x = age_group, 
             y = n)) +
  geom_col(fill = 'blue') +
  labs(title = "Age Distribution",
       x = "Age Group",
       y = "Count") +
  theme(plot.title = element_text(hjust = 0.5))
age_distribution
```

## Distribution of Cancer Types According to Age

```{r}

cancertype_age <-
  clinical_data |> 
  mutate(age_group = cut(age_at_diagnosis, 
                         breaks = c(20, 30, 40, 50, 60, 70, 80, 100))) |> 
  mutate(age_group = factor(age_group),
         cancer_type_detailed = factor(cancer_type_detailed)) |> 
  filter(!is.na(cancer_type_detailed)) |> 
  group_by(age_group, 
           cancer_type_detailed) |> 
  summarise(n = n()) |>
  mutate(percentage = n / sum(n) * 100) |> 
  ggplot(aes(x = age_group, 
             y = percentage , 
             fill = cancer_type_detailed)) +
  geom_col(position = "dodge",
           alpha = 0.8) +
  labs(title = "Distribution of Cancer Types According to Age",
       x = "Age Group",
       y = "Percentage(%)",
       fill = "Cancer Type") +
  theme(legend.text = element_text(size = 6),
        legend.title = element_text(size = 8),
        legend.position = 'bottom',
        plot.title = element_text(hjust = 0.5))

cancertype_age
```

# Saving plots as .png

```{r}
ggsave("03_cancertype_age.png", 
       path = "../doc/images/",
       plot = cancertype_age,
       width = 8,
       height = 5)

ggsave("03_age_distribution.png", 
       path = "../doc/images/",
       plot = age_distribution,
       width = 8,
       height = 5)
```
