==> R/05_analysis_survival.qmd <==
---
title: "05_analysis_survival"
format: html
editor: visual
---

# Load data

```{r}
clinical_data <- 
  read_tsv(
  file = "../data/clinical_data_aug.tsv")
```

```{r}

clinical_data <-
  na.omit(clinical_data)

```

# Survival analysis

With a survival analysis of the breast cancer dataset, we will visualize survival curves to understand the overall survival probability over time for breast cancer patients. Evaluate the importance of key variables such as age, tumor size, mutation number and treatment options on overall survival. Furthermore, we will identify factors that significantly influence survival outcomes for breast cancer patients and compare survival curves between patient subgroups based on specific characteristics, such as ER status and treatment type.

## Survival Rate

By calculating the mean value of the patient survival, the value for 'survival_rate' will represent the average survival rate.

```{r}

survival_rate <- 
  clinical_data |>
  summarise(survival_mean = mean(overall_survival))

print(survival_rate)

```

A survival rate of 0.43 means that, on average 43% of the individuals in the data set have survived the breast cancer.

## Age Distribution

Converting the variable 'overall_survival' to a factor is done to represent it as a categorical variable, and to avoid misinterpretation of the numerical values given in the variable.

To explore the distribution of ages at diagnosis for patients who survived and those who did not, we identified whether certain age at diagnosis are associated with survived and died patients:

```{r}
age_boxplot <- 
  ggplot(data = clinical_data,
         mapping = aes(x = overall_survival, 
                       y = age_at_diagnosis, 
                       fill = factor(overall_survival))) +
  geom_boxplot(alpha = 0.5, 
               show.legend = FALSE) +
  labs(title = "The distribution of Patient Survival by Ages at Diagnosis",
       x = "Patient Survival", 
       y = "Age at Diagnosis", 
       subtitle = str_c("Labelling: Died = 0, ",
                        "Survived = 1")) +
  scale_fill_manual(values = c("0" = "red",
                               "1" = "green")) +
  theme_classic(base_size = 12,
                base_family = "Avenir")

age_boxplot
```

To observe whether certain age groups have higher proportion of survival, we visualize the distribution of age groups based on the patient survival status:

```{r}

age_bar <- 
  clinical_data |>
  ggplot(mapping = aes(x = age_group,
                       fill = factor(overall_survival))) +
  geom_bar(alpha = 0.5,
           position = "dodge") +
  labs(title = "Distribution of Age Groups in terms of Patient Survival",
       x = "Age Groups",
       y = "Number of patients",
       fill = "Patient Survival") +
  scale_fill_manual(values = c("0" = "red",
                               "1" = "green"), 
                    labels = c("Died",
                               "Survived")) +
  theme_classic(base_size = 12,
                base_family = "Avenir")
age_bar
```

Each bar represents an age group, and the bars are colored according to survival status. The height of each bar indicated the number of patients in each age group.

The box plots below illustrate a distribution of tumor diameter (mm) and survival of the patients across different age groups. Each box represents a specific age group, and the vertical axis shows the tumor diameter. The bars are colored according to survival status.

```{r}

age_tumor_survival_boxplot <- 
  ggplot(data = clinical_data, 
         mapping = aes(x = age_group, 
                       y = tumor_size, 
                       fill = factor(overall_survival))) +
  geom_boxplot(alpha = 0.5, 
               show.legend = TRUE) +
  labs(title = "The Distribution of Tumor Size and Overall Survival by Age Groups",
       x = "Age Group (years)",
       y = "Tumor Diameter (mm)",
       fill = "Patient Survival") +
  scale_fill_manual(values = c("0" = "red",
                               "1" = "green"),
                    labels = c("Died",
                               "Survived")) +
  theme_classic(base_size = 12,
                base_family = "Avenir")


age_tumor_survival_boxplot
```

The variability of tumor sizes within each age group differs less for each age groups. by looking at the points outside the whiskers, many more outliers are represent for the age group 41-60 and 61-80, which means more unusal tumor sizes are represent within the age groups compared to the age groups 0-40 and 81+. Furthermore, it is clear to observe from the whiskers for each age group that died patients had a larger tumor size compared to tumor sizes for survived patients.

## The Impact of Tumor Characteristics on Patient Survival

To understand how specific attributes of tumors influence the survival outcomes of patients, we made some relevant plot to create visualizations to effectively communicate the results.

We have created box plots of positive lymph nodes based on logarithmic transformed tumor sizes.

```{r}

lymph_boxplot <- 
  clinical_data |> 
  ggplot(aes(y = as.factor(lymph_nodes_examined_positive),
             x = tumor_size_log, 
             fill = factor(overall_survival))) +
  geom_boxplot(alpha = 0.5) +
  labs(title = "The Distribution of Positive Lymph Nodes according to log-transformed Tumor Sizes",
       x = "Log10 Transformed Tumor Size", 
       y = "Positive Lymph Nodes",
       fill = "Overall Survival") +
  theme_classic(base_size = 9,
                base_family = "Avenir") +
  theme(legend.position = "bottom")

lymph_boxplot
```

The box plot shows the average tendency of log10 transofrmed tumor sizes for different levels of positive lymph nodes. The interquartile range of each box indicated the variability of log-transformed tumor sizes within each level. so the larger boxes shows greater variability.

## Clinical Treatment traits

In order to visualize the distribution of the patient survival status based on different treatment traits, such as: chemotherapy, hormone therapy and radio therapy, three bar chart plots are created below.

```{r}

chemotherapy_chart <-
  clinical_data |>
  ggplot(aes(x = factor(chemotherapy), 
             fill = factor(overall_survival))) +
  geom_bar(alpha = 0.5,
           position = "dodge") +
  labs(title = "The Survival of Breast Cancer Patients Treated with Different Clinical Treatments",
       x = "Chemotherapy",
       y = "Number of Patients",
       fill = "Survival") +
  scale_fill_manual(values = c("0" = "red",
                               "1" = "green"), 
                    labels = c("Died", "Survived")) +  
  scale_x_discrete(labels = c("0" = "No",
                              "1" = "Yes")) +
  theme_classic(base_size = 9,
                base_family = "Avenir")


hormone_chart <-
  clinical_data |>
  ggplot(aes(x = factor(hormone_therapy), 
             fill = factor(overall_survival))) +
  geom_bar(alpha = 0.5,
           position = "dodge") +
  labs(x = "Hormone Therapy",
       y = "Number of Patients",
       fill = "Survival") +
  scale_fill_manual(values = c("0" = "red",
                               "1" = "green"), 
                    labels = c("Died",
                               "Survived")) +  
  scale_x_discrete(labels = c("0" = "No",
                              "1" = "Yes")) +
  theme_classic(base_size = 10,
                base_family = "Avenir")



radio_chart <- 
  clinical_data |>
  ggplot(aes(x = factor(radio_therapy), 
             fill = factor(overall_survival))) +
  geom_bar(alpha = 0.5,
           position = "dodge") +
  labs(x = "Radio Therapy",
       y = "Number of Patients",
       fill = "Survival") +
  scale_fill_manual(values = c("0" = "red",
                               "1" = "green"), 
                    labels = c("Died",
                               "Survived")) +  
  scale_x_discrete(labels = c("0" = "No",
                              "1" = "Yes")) +
  theme_classic(base_size = 10,
                base_family = "Avenir")

plot_surv_treatment <- chemotherapy_chart/(hormone_chart + radio_chart)
plot_surv_treatment
```

The height of each bar represents the number of patients, and the color indicates their survival status

## Survival efficacy related to different types of surgery

```{r}
p1 <- 
  clinical_data |> 
  filter(type_of_breast_surgery == 'MASTECTOMY') |>  
  filter(death_from_cancer!='Died of Other Causes')|> 
  ggplot(aes(x = death_from_cancer, 
             fill = death_from_cancer)) + 
  geom_bar(alpha = 0.7) + 
  theme_minimal() + 
  scale_fill_manual(values = c("Died of Disease" = "red",
                               "Living" = "green"), 
                    labels = c("Died",
                               "Survived")) +
  labs (x ='masectomy',
        y='number of individuals')+ 
  theme(axis.title.x = element_text(size = 12,
                                    face = "bold"), 
        axis.title.y = element_text(size = 12,
                                    face = "bold"), 
        axis.text.x = element_text(size = 10), 
        axis.text.y = element_text(size = 10)) + 
  theme(legend.position = "none") 

  
p2 <- 
  clinical_data |> 
  filter(type_of_breast_surgery == 'BREAST CONSERVING') |>  
  filter(death_from_cancer!= 'NA' & death_from_cancer!='Died of Other Causes') |> 
  ggplot(aes(x = death_from_cancer,
             fill = death_from_cancer)) + 
  geom_bar(alpha = 0.7) + 
  theme_minimal() + 
  scale_fill_manual(values = c("Died of Disease" = "red",
                               "Living" = "green"), 
                    labels = c("Died of Disease",
                               "Survived")) +
  labs (x ='breast conserving',
        y='number of individuals') + 
  theme(axis.title.x = element_text(size = 12,
                                    face = "bold"), 
        axis.title.y = element_text(size = 12,
                                    face = "bold"), 
        axis.text.x = element_text(size = 10), 
        axis.text.y = element_blank()) + 
  labs(y=NULL)


plot_surv_surg <- p1 +
  labs(title = "Survival Efficacy Related to Different Types of Surgery") +
  p2 

plot_surv_surg
```

## ER status and overall survival

```{r}
plot_er_status <-   
  clinical_data |> 
  ggplot(aes(x = er_status,
             fill = factor(overall_survival))) +
  geom_bar(alpha = 0.7, 
           position = "dodge") +
  labs(x = "ER status",
       y = "Number of Patients",
       fill = "Overall Survival")


plot_pr_status <-   
  clinical_data |> 
  ggplot(aes(x = pr_status,
             fill = factor(overall_survival))) +
  geom_bar(alpha = 0.7, 
           position = "dodge") +
  labs(x = "PR status",
       y = "Number of Patients",
       fill = "Overall Survival")

plot_her2_status <-   
  clinical_data |> 
  ggplot(aes(x = her2_status,
             fill = factor(overall_survival))) +
  geom_bar(alpha = 0.7, 
           position = "dodge") +
  labs(x = "HER2 status",
       y = "Number of Patients",
       fill = "Overall Survival")

plot_hormone_status_boxplot <- plot_er_status / (plot_pr_status + plot_her2_status)
plot_hormone_status_boxplot
```

## Neoplasm Histological Grade vs overall survival

```{r}
violin_plot <- 
  clinical_data |> 
  ggplot(aes(x = factor(neoplasm_histologic_grade), 
             y = overall_survival_months, 
             fill = factor(overall_survival))) +
  geom_boxplot() +
  labs(title = "Neoplasm Histologic Grade and Overall Survival",
       x = "Neoplasm Histologic Grade",
       y = "Overall survival months",
       fill = "Overall survival") +
  theme_classic(base_size = 10,
                base_family = "Avenir")

violin_plot
```

## Therapeutic Efficacy Comparison

```{r}

living <- 
  clinical_data |> 
  filter(death_from_cancer == "Living")
n_living = nrow(living)

death_cancer <- clinical_data |> 
  filter(death_from_cancer == "Died of Disease")
n_death_cancer = nrow(death_cancer)

n_tot = n_living # + n_death_cancer

therapy <- 
  living |> 
  summarise(
    n_C = sum(chemotherapy == "1" & 
                radio_therapy == "0" & hormone_therapy == "0" ),
    
    n_R = sum(radio_therapy == "1" & 
                chemotherapy == "0" & hormone_therapy == "0"),
    
    n_H = sum(hormone_therapy == "1" & 
                chemotherapy == "0" & radio_therapy == "0" ),
    
    n_CR = sum(chemotherapy == "1" & radio_therapy == "1" & 
                 hormone_therapy == "0" ),
    
    n_CH = sum(chemotherapy == "1" & hormone_therapy == "1" & 
                 radio_therapy == "0"),
    
    n_RH = sum(radio_therapy == "1" & hormone_therapy == "1" & 
                 chemotherapy == "0" ),
    
    n_RHC = sum(chemotherapy == "1" & radio_therapy == "1" & hormone_therapy == "1")
  ) |> 
  mutate(
    perc_n_C = round(n_C / n_tot * 100, 2),
    perc_n_R = round(n_R / n_tot * 100, 2),
    perc_n_H = round(n_H / n_tot * 100, 2),
    perc_n_CR = round(n_CR / n_tot * 100, 2),
    perc_n_CH = round(n_CH / n_tot * 100, 2),
    perc_n_RH = round(n_RH / n_tot * 100, 2),
    perc_n_RHC = round(n_RHC / n_tot * 100, 2)
  )


therapy_data_long <- 
  therapy |> 
  pivot_longer(
    cols = starts_with("perc_n"),
               names_to = "Category",
               values_to = "Var1"
    )


plot_therapeutic <-
  ggplot(therapy_data_long,
         aes(x = factor(Category,
                        levels = Category),
             y = Var1,
             fill = Category)) +
  geom_bar(stat = "identity",
           position = "dodge") +
  theme(axis.text.x = element_text(angle = 45,
                                   hjust = 1),
        legend.position = "none") +
  labs(x = "Therapy",
       y = "%",
       fill = "Therapy") +
  labs(title = "Therapeutic Efficacy Comparison: 
       A Study on Therapies Used by Individuals Alive") +
  scale_x_discrete(labels = c(
    perc_n_C = "Chemo",
    perc_n_R = "Radio",
    perc_n_H = "Hormone",
    perc_n_CR = "Chemo + Radio",
    perc_n_CH = "Chemo + Hormone",
    perc_n_RH = "Radio + Hormone",
    perc_n_RHC = "Chemo + Radio + Hormone"
  ))

plot_therapeutic
```

# Write images

```{r}
ggsave("05_age_boxplot.png", 
       path = "../doc/images/",
       plot = age_boxplot,
       width = 8,
       height = 5)

ggsave("05_age_bar.png", 
       path = "../doc/images/",
       plot = age_bar,
       width = 8,
       height = 5)

ggsave("05_age_tumor_survival_boxplot.png", 
       path = "../doc/images/",
       plot = age_tumor_survival_boxplot,
       width = 8,
       height = 5)

ggsave("05_lymph_boxplot.png", 
       path = "../doc/images/",
       plot = lymph_boxplot,
       width = 8,
       height = 5)

ggsave("05_surv_treatment.png", 
       path = "../doc/images/",
       plot = plot_surv_treatment,
       width = 8,
       height = 5)

ggsave("05_surv_surg.png", 
       path = "../doc/images/",
       plot = plot_surv_surg,
       width = 8,
       height = 5)

ggsave("05_violin_plot.png", 
       path = "../doc/images/",
       plot = violin_plot,
       width = 8,
       height = 5)

ggsave("05_hormone_status_boxplot.png", 
       path = "../doc/images/",
       plot = plot_hormone_status_boxplot,
       width = 8,
       height = 5)

ggsave("05_therapeutics_comp.png", 
       path = "../doc/images/",
       plot = plot_therapeutic,
       width = 8,
       height = 5)
```
